requirements.txt
python-telegram-bot[asyncio]==20.8
SQLAlchemy[asyncio]==2.0.27
aiosqlite==0.19.0
Pillow==10.2.0
python-dotenv==1.0.1
